add AddLayer() method
